pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mirrorbox"
version = "2.0.0"
authors = [
    { name = "Pouya Rezapour", email = "pouyarezapourr@gmail.com" },
]
description = "Enterprise Docker Mirror Client with AI-driven optimization."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT License" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
    "typer",
    "rich",
    "requests",
    "pyyaml",
    "pywebview", 
    "platformdirs"
]

[project.urls]
"Homepage" = "https://github.com/pouyarer/mirrorbox"
"Repository" = "https://github.com/pouyarer/mirrorbox"

[project.scripts]
mirrorbox = "mirrorbox.cli:app"

[tool.setuptools.packages.find]
include = ["mirrorbox"]